Split WHOLE codebase into separate files
