+++ .gitignore
@@ -88,6 +88,9 @@ venv.bak/
 # pytype static type analyzer
 .pytype/
 
+# vscode
+.vscode/
+
 # Cython debug symbols
 cython_debug/
 
@@ -109,3 +112,4 @@ logs/oui-lookup/*.txt
 logs/port-scanner/*.txt
 
 logs/vuln-scanner/*.txt
+
