+++ functions/utils/exit_program.py
@@ -1,0 +1,6 @@
+import sys
+from functions.utils.font_styles import *
+
+# Message to be printed when using exit command or Ctrl-C inputted
+def exit():
+    sys.exit(success_message('Shutting down NetSploit ... Bye! ( ^_^)/'))
